Cassandra DataStax Astra Database created and connected

+++ MeteorologicalDataApp/src/main/java/com/korinek/MeteorologicalDataApp/service/MeasurementService.java
@@ -3,14 +3,12 @@ package com.korinek.MeteorologicalDataApp.service;
 import com.korinek.MeteorologicalDataApp.model.Measurement;
 import com.korinek.MeteorologicalDataApp.repository.MeasurementRepository;
 import jakarta.persistence.EntityNotFoundException;
-import org.hibernate.cfg.NotYetImplementedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -25,12 +23,11 @@ public class MeasurementService {
     }
 
     public void addNewMeasurement(Measurement measurement) {
-        /*if(this.measurementRepository.existsById(measurement.getId()) || this.measurementRepository.existsByName(measurement.getName())){
+        if(this.measurementRepository.existsById(measurement.getId())){
             throw new DuplicateKeyException("Already exists!");
         } else {
             this.measurementRepository.save(measurement);
-        }*/
-        throw new NotYetImplementedException();
+        }
     }
 
     public void addNewMeasurements(List<Measurement> measurements) {
@@ -40,46 +37,42 @@ public class MeasurementService {
     }
 
     public Measurement getMeasurement(int id) {
-        /*if(this.measurementRepository.existsById(id)){
-            Optional<Measurement> measurementOptional = this.measurementRepository.findById(id);
+        if(this.measurementRepository.existsById(id)){
+            return this.measurementRepository.findById(id);
+            /*Optional<Measurement> measurementOptional = this.measurementRepository.findById(id);
             if (measurementOptional.isPresent()) {
                 return measurementOptional.get();
             } else {
                 throw new EntityNotFoundException();
-            }
+            }*/
         } else {
             throw new EntityNotFoundException();
-        }*/
-        throw new NotYetImplementedException();
+        }
     }
 
     @Transactional(readOnly = true)
     public List<Measurement> getAllMeasurements() {
-        //return StreamSupport.stream(measurementRepository.findAll().spliterator(), false).collect(Collectors.toList());
-        throw new NotYetImplementedException();
+        return StreamSupport.stream(measurementRepository.findAll().spliterator(), false).collect(Collectors.toList());
     }
 
     public void updateMeasurement(int id, Measurement measurement) {
-        /*if(this.measurementRepository.existsById(id)){
+        if(this.measurementRepository.existsById(id)){
             measurement.setId(id);
             this.measurementRepository.save(measurement);
         } else {
             throw new EntityNotFoundException();
-        }*/
-        throw new NotYetImplementedException();
+        }
     }
 
     public void deleteMeasurement(int id) {
-        /*if(this.measurementRepository.existsById(id)){
+        if(this.measurementRepository.existsById(id)){
             this.measurementRepository.deleteById(id);
         } else {
             throw new EntityNotFoundException();
-        }*/
-        throw new NotYetImplementedException();
+        }
     }
 
     public void deleteAllMeasurements() {
-        //this.measurementRepository.deleteAll();
-        throw new NotYetImplementedException();
+        this.measurementRepository.deleteAll();
     }
 }
